requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.40.0
psycopg2-binary==2.9.9
schedule==1.2.0
python-dotenv==1.0.0
fake-useragent==1.4.0
aiohttp==3.9.1
asyncio==3.4.3
